+++ packages/x-charts/src/BarChart/BarLabel/index.ts
@@ -2,5 +2,5 @@ export { BarLabel } from './BarLabel';
 export type { BarLabelProps } from './BarLabel';
 export { barLabelClasses, getBarLabelUtilityClass } from './barLabelClasses';
 export type { BarLabelSlotProps, BarLabelSlots } from './BarLabelItem';
-export type { BarLabelOwnerState, BarItem, BarLabelContext } from './types';
+export type { BarLabelOwnerState, BarItem, BarLabelContext } from './BarLabel.types';
 export type { BarLabelClasses, BarLabelClassKey } from './barLabelClasses';
